requirements.txt
torch
torchvision
fastapi[standard]
uvicorn[standard]
python-multipart
pillow
numpy
matplotlib
tqdm
tensorboard
